move and delete obsolete files to clean up repository
