+++ pyproject.toml
@@ -1,8 +1,74 @@
 [build-system]
-requires = ["flit_core >=3.2,<4"]
-build-backend = "flit_core.buildapi"
+requires = ["setuptools>=42"]
+build-backend = "setuptools.build_meta"
 
 [project]
-name = "lumache"
-authors = [{name = "Graziella", email = "graziella@lumache"}]
-dynamic = ["version", "description"]
+name = "pystructurePipeline"
+version = "4.3.0"
+description = "Homogenize and analyze multi-wavelength astronomical datasets on hexagonal grids."
+readme = "README.md"
+license = { file = "LICENSE" }
+authors = [
+    { name = "Jakob den Brok",  email = "jadenbrok@mpia.de"     },
+    { name = "Lukas Neumann",   email = "lukas.neumann@eso.org" },
+]
+keywords = ["astronomy", "astrophysics", "FITS", "radio", "pipeline"]
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Intended Audience :: Science/Research",
+    "Topic :: Scientific/Engineering :: Astronomy",
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+]
+requires-python = ">=3.8"
+dependencies = [
+    "astropy>=5.0",
+    "numpy>=1.22",
+    "pandas>=1.4",
+    "scipy>=1.7",
+    "matplotlib>=3.4",
+    "reproject>=0.9",
+    "radio_beam>=0.3.4",
+    "spectral_cube>=0.6",
+]
+
+[project.optional-dependencies]
+dev = [
+    "pytest>=7.0",
+    "pytest-cov",
+    "ruff",
+    "black",
+]
+docs = [
+    "sphinx",
+    "sphinx-rtd-theme",
+    "numpydoc",
+]
+
+[project.scripts]
+pystructure = "pystructurePipeline.cli:main"
+
+[project.urls]
+Homepage      = "https://github.com/PhangsTeam/PyStructure"
+Repository    = "https://github.com/lukas-neumann-astro/PyStructure"
+Issues        = "https://github.com/PhangsTeam/PyStructure/issues"
+Documentation = "https://pystructure.readthedocs.io/en/latest/"
+
+[tool.setuptools.packages.find]
+where   = ["."]
+include = ["pystructurePipeline*"]
+
+[tool.setuptools.package-data]
+pystructurePipeline = [
+    "templates/keys/*.txt",
+    "templates/run_pystructure.py",
+    "py.typed",
+]
+
+[tool.pytest.ini_options]
+testpaths = ["pystructurePipeline"]
+python_files = ["test_*.py"]
