--- scripts/sampling_at_resol.py
@@ -1,440 +1,0 @@
-import numpy as np
-import matplotlib.pyplot as plt
-from astropy.io import fits
-# import os.path
-from os import path
-import copy
-import radio_beam # LN: not used in the script
-from scipy import stats # LN: not used in the script
-from astropy import units as u # LN: not used in the script
-from reproject import reproject_interp
-from astropy.wcs import WCS
-from astropy.stats import median_absolute_deviation
-from astropy.convolution import Gaussian1DKernel, Box1DKernel, convolve
-# from spectral_cube import SpectralCube # LN: not used in the script
-from twod_header import twod_head
-import warnings
-warnings.filterwarnings("ignore")
-
-
-from gauss_conv import *
-
-
-def get_vaxis(hdr):
-    v = np.arange(hdr["NAXIS3"])
-    vdif = v-(hdr["CRPIX3"]-1)
-    vaxis = vdif * hdr["CDELT3"] + hdr["CRVAL3"]
-    return vaxis
-
-
-def sample_at_res(in_data,
-                  ra_samp,
-                  dec_samp,
-                  in_hdr = None,
-                  target_res_as = None,
-                  target_hdr = None,
-                  coverage = None,
-                  rms = None,  # LN: not used in the function (delete?)
-                  show = False,  # LN: not used in the function (delete?)
-                  #advanced parameters
-                  line_name = "",
-                  source = "",
-                  save_fits = False,
-                  path_save_fits = "",
-                  perbeam = False,
-                  spec_smooth = ["default","binned"],
-                  unc=False):
-    """
-    Function to sample the data and convolve
-    """
-
-    #--------------------------------------------------------------
-    #   Defaults and Definitions
-    #-------------------------------------------------------------
-
-    # Check if ra and dec arrays have same length
-    if len(ra_samp) != len(dec_samp):
-        print(f'{"[ERROR]":<10}', 'Need matching RA and Dec vector. Returning.')
-        return np.nan
-
-    # Check if data given as string (need to load data first) or already given
-    # as an array.
-    if isinstance(in_data, str):
-        # Data given in form of a string, need to load data
-        if not path.exists(in_data):
-            print(f'{"[ERROR]":<10}', f'File {in_data} not found. Returning.')
-            return ra_samp * np.nan
-        data, hdr = fits.getdata(in_data, header = True)
-
-    else:
-        data = copy.deepcopy(in_data)
-        if not in_hdr is None:
-            hdr = in_hdr
-        else:
-            print(f'{"[ERROR]":<10}', 'Provide a valid header with the data.')
-            return ra_samp*np.nan
-
-    # Check if target resolution was provided
-    if target_res_as is None:
-        print(f'{"[WARNING]":<10}', 'No target resolution specified.\n \
-              Taking this as zero and sampling at the native resolution.')
-        target_res_as = 0
-
-    if target_hdr is None:
-        print(f'{"[WARNING]":<10}', 'No target header. Will use native astrometry.')
-
-
-    dim_data = np.shape(data)
-    if len(dim_data) == 3:
-        is_cube = True
-    else:
-        is_cube = False
-
-    if not is_cube:
-        target_hdr = twod_head(target_hdr)
-
-    #--------------------------------------------------------------
-    #   Convolve and Align
-    #------------------------------------------------------------
-
-    # skip convolution, if resolution is not provided in the header
-    if 'BMAJ' not in hdr:
-        print(f'{"[WARNING]":<10}', f'Header does not contain beam information (missing key: BMAJ).')
-        print(f'{"[INFO]":<10}', f'Skip convolution and sample.')
-        hdr_out = copy.copy(hdr)
-    else:
-        current_bmaj = hdr["BMAJ"]
-
-        #reduce by 1% to account for rounding
-        if current_bmaj < (0.99*target_res_as/3600):
-            
-            print(f'{"[INFO]":<10}', f'Convolve to {np.round(target_res_as, 3)} arcseconds.')
-
-            # convolve data with Gaussian kernel
-            data, hdr_out = conv_with_gauss(in_data = data, 
-                                            in_hdr = hdr,
-                                            target_beam = target_res_as*np.array([1,1,0]),
-                                            quiet = True,
-                                            perbeam = perbeam,
-                                            unc=unc)
-
-            #data_speccube = SpectralCube(data=data, wcs = WCS(hdr))
-            #data_speccube.beam = radio_beam.Beam(major=current_bmaj*u.arcsec, minor=current_bmaj*u.arcsec, pa=0*u.deg)
-            #beam = radio_beam.Beam(major=target_res_as*u.arcsec, minor=target_res_as*u.arcsec, pa=0*u.deg)
-            #new_datacube = data_speccube.convolve_to(beam)
-            #data = np.array(new_datacube.unmasked_data[:,:,:])
-            #hdr_out = hdr
-        else:
-            print(f'{"[INFO]":<10}', 'Already at target resolution.')
-            print(f'{"[INFO]":<10}', f'Skip convolution and sample.')
-            hdr_out = copy.copy(hdr)
-
-    #; Measure the rms in the map after convolution but before
-    #; alignment. This isn't really ideal but can be a useful
-    #; shorthand for the noise.
-    rms = median_absolute_deviation(data, axis = None,ignore_nan=True) # LN: not used in the function (delete?)
-
-    ######
-    # Check that velocity axis is in m/s, not km/s
-    ######
-    if is_cube:
-        if abs(target_hdr["CDELT3"])<200:
-            print(f'{"[INFO]":<10}', 'Overlay cube in km/s, converting to m/s.')
-            target_hdr["CDELT3"] = 1000 * target_hdr["CDELT3"]
-            target_hdr["CRVAL3"] = 1000 * target_hdr["CRVAL3"]
-            target_hdr["CUNIT3"] = "m/s"
-
-        if abs(hdr["CDELT3"])<200:
-            print(f'{"[INFO]":<10}', 'Line cube in km/s, converting to m/s.')
-            hdr_out["CDELT3"] = 1000 * hdr["CDELT3"]
-            hdr_out["CRVAL3"] = 1000 * hdr["CRVAL3"]
-            hdr_out["CUNIT3"] = "m/s"
-
-        #check if cube vaxis is inverted (want delta v>0):
-        if target_hdr["CDELT3"]<0:
-            print(f'{"[INFO]":<10}', 'Target cube has invertied vaxis. Re-inverting.')
-            vaxis_inv = get_vaxis(target_hdr)
-            target_hdr["CDELT3"] = -1* target_hdr["CDELT3"]
-            target_hdr["CRPIX3"] = 1
-            target_hdr["CRVAL3"] = vaxis_inv[-1]
-
-        if hdr["CDELT3"]<0:
-            print(f'{"[INFO]":<10}', 'Line cube has invertied vaxis. Re-inverting.')
-            vaxis_inv = get_vaxis(hdr_out)
-            hdr_out["CDELT3"] = -1* hdr["CDELT3"]
-            hdr_out["CRPIX3"] = 1
-            hdr_out["CRVAL3"] = vaxis_inv[-1]
-
-            #flip the cube
-            data = np.flip(data, axis=0)
-
-    #perform spectral smooting, if needed
-    if spec_smooth[0] in ["overlay"]:
-        spec_smooth[0] = abs(target_hdr["CDELT3"])/1000
-
-    if type(spec_smooth[0]) == int or type(spec_smooth[0]) == float:
-        spec_res = abs(hdr["CDELT3"])/1000
-        fwhm_factor = np.sqrt(8*np.log(2))
-        if spec_res >= spec_smooth[0]:
-            print(f'{"[INFO]":<10}', 'No spectral smoothing; already at target resolution.')
-        else:
-            print(f'{"[INFO]":<10}', f'Do spectral smoothing to {np.round(spec_smooth[0], 3)} km/s.')
-
-
-            #check the method:
-            #
-            # Gauss method:
-            if spec_smooth[1] in ["gauss"]:
-                print(f'{"[INFO]":<10}', f'Do spectral smoothing via Gaussian convolution.')
-                pix = ((spec_smooth[0]**2 - spec_res**2)**0.5  / spec_res)/fwhm_factor
-                kernel = Gaussian1DKernel(pix)
-
-                for spec_n in ProgressBar(range(dim_data[1]*dim_data[2])):
-                    y = spec_n%dim_data[1]
-                    x = spec_n//dim_data[1]
-                    data[:,y,x] = convolve(data[:, y,x],kernel,nan_treatment='fill')
-
-            # Binning method
-            elif spec_smooth[1] in ["binned", "combined"]:
-                vaxis_native = get_vaxis(hdr_out)
-
-                n_ratio = int(spec_smooth[0]/spec_res)
-
-                #smooth, if remainder of target/res is almost 90%
-                if (spec_smooth[0]/spec_res-n_ratio)>0.9:
-                    n_ratio+=1
-                new_len = len(vaxis_native)//n_ratio
-                if n_ratio==1:
-                    print(f'{"[INFO]":<10}', 'No spectral binning possible; already at target resolution.')
-                else:
-                    print(f'{"[INFO]":<10}', f'Do spectral smoothing via channel binning.')
-                    new_vaxis = np.array([np.nanmean(vaxis_native[n_ratio*j:n_ratio*(j+1)]) for j in range(new_len)])
-                    data = np.array([np.nanmean(data[n_ratio*j:n_ratio*(j+1),:,:], axis=0) for j in range(new_len)])
-
-                    hdr_out["NAXIS3"] = new_len
-                    hdr_out["CDELT3"] = new_vaxis[1]-new_vaxis[0]
-                    hdr_out["CRVAL3"] = new_vaxis[0] + (hdr_out["CRPIX3"]-1)*hdr_out["CDELT3"]
-
-                if spec_smooth[1] in ["combined"]:
-                    print(f'{"[INFO]":<10}', f'Do spectral smoothing via Gaussian convolution.')
-                    if n_ratio*spec_res<spec_smooth[0]:
-                        pix = ((spec_smooth[0]**2 - (n_ratio*spec_res)**2)**0.5  / spec_res)/fwhm_factor
-                        kernel = Gaussian1DKernel(pix)
-
-
-                        for spec_n in ProgressBar(range(dim_data[1]*dim_data[2])):
-                            y = spec_n%dim_data[1]
-                            x = spec_n//dim_data[1]
-                            data[:,y,x] = convolve(data[:, y,x],kernel,nan_treatment='fill')
-
-    # Align, if needed
-    if not target_hdr is None:
-        #check if we are spectrally smoothing. If so, prep the target_hdr to the right spectral number
-        trg_hdr = copy.deepcopy(target_hdr)
-        if not spec_smooth[0] in ["default"]:
-
-            #check if we need to spectrally change the target resolution
-            if spec_smooth[0] > trg_hdr["CDELT3"]/1000:
-                vaxis_ov = get_vaxis(trg_hdr)
-                new_vaxis = np.arange(vaxis_ov[0], vaxis_ov[-1],spec_smooth[0]*1000)
-
-                trg_hdr["NAXIS3"] = len(new_vaxis)
-                trg_hdr["CDELT3"] = spec_smooth[0]*1000
-                trg_hdr["CRVAL3"] = new_vaxis[0] + (trg_hdr["CRPIX3"]-1)*trg_hdr["CDELT3"]
-
-        wcs_target = WCS(trg_hdr)
-        #need to delete the rest frequency for latest astropy version
-        del hdr_out["RESTF*"]
-        del trg_hdr["RESTF*"]
-
-        data_out, footprint = reproject_interp((data,hdr_out), trg_hdr, order="nearest-neighbor")
-        data = data_out
-
-        # Save the convloved file as a fits file
-        if save_fits:
-            out_header = copy.copy(trg_hdr)
-            out_header["BMAJ"]=target_res_as/3600
-            out_header["BMIN"]=target_res_as/3600
-            out_header["LINE"]=line_name
-            fits.writeto(path_save_fits+source+'_'+str(line_name)+'_{}as.fits'.format(target_res_as), data=data_out, header=out_header, overwrite=True)
-            print(f'{"[INFO]":<10}', 'Convolved fits file has been saved.')
-
-    else:
-        print(f'{"[INFO]":<10}', 'No alignment because no target header supplied.')
-
-    #--------------------------------------------------------------
-    #   Sample
-    #--------------------------------------------------------------
-
-    wcs_target = WCS(trg_hdr)
-    if is_cube:
-        pixel_coords = wcs_target.all_world2pix(np.column_stack((ra_samp, dec_samp, np.zeros(len(dec_samp)))),0)
-    else:
-        pixel_coords = wcs_target.all_world2pix(np.column_stack((ra_samp, dec_samp)),0)
-    samp_x = np.array(np.rint(pixel_coords[:,0]),dtype=int)
-    samp_y = np.array(np.rint(pixel_coords[:,1]),dtype= int)
-
-
-
-    n_pts = len(samp_x)
-    dim_data = np.shape(data)
-
-
-    if is_cube:
-        result = np.zeros((n_pts, dim_data[0]))*np.nan
-    else:
-        result = np.zeros(n_pts)*np.nan
-
-    coverage = np.zeros(n_pts)
-    if is_cube:
-
-        in_map = np.where((samp_x>0)& (samp_x <dim_data[2]) &
-                          (samp_y > 0)& (samp_y < dim_data[1]))[0]
-        in_map_ct = len(in_map)
-    else:
-        in_map = np.where((samp_x>0)& (samp_x <dim_data[1]) &
-                          (samp_y > 0)& (samp_y < dim_data[0]))[0]
-        in_map_ct = len(in_map)
-
-
-    if in_map_ct>0:
-        if is_cube:
-            for kk in range(in_map_ct):
-
-                result[in_map[kk],:] = data[:, samp_y[in_map[kk]],samp_x[in_map[kk]]]
-                coverage[in_map[kk]] =  np.nansum(np.isfinite(data[:, samp_y[in_map[kk]],samp_x[in_map[kk]]]))>=1
-            coverage = np.array(coverage, dtype = int)
-        else:
-            result[in_map] = data[samp_y[in_map],samp_x[in_map]]
-            coverage[in_map] = np.isfinite(data[samp_y[in_map],samp_x[in_map]])
-
-    #--------------------------------------------------------------
-    #   Return
-    #------------------------------------------------------------
-
-    return result, trg_hdr
-
-
-def sample_mask(in_data,
-                  ra_samp,
-                  dec_samp,
-                  in_hdr = None,
-                #   target_res_as = None,
-                  target_hdr = None,
-                  coverage = None,
-                #   rms = None,  # LN: not used in the function (delete?)
-                #   show = False,  # LN: not used in the function (delete?)
-                  #advanced parameters
-                #   line_name = "",
-                #   source = "",
-                #   save_fits = False,
-                #   path_save_fits = "",
-                #   perbeam = False,
-                #   spec_smooth = ["default","binned"],
-                #   unc=False
-                  ):
-
-    """
-    Function to sample the data
-    """
-
-    #--------------------------------------------------------------
-    #   Defaults and Definitions
-    #-------------------------------------------------------------
-
-    # Check if ra and dec arrays have same length
-    if len(ra_samp) != len(dec_samp):
-        print(f'{"[ERROR]":<10}', 'Need matching RA and Dec vector. Returning.')
-        return np.nan
-
-    # Check if data given as string (need to load data first) or already given
-    # as an array.
-    if isinstance(in_data, str):
-        # Data given in form of a string, need to load data
-        if not path.exists(in_data):
-            print(f'{"[ERROR]":<10}', f'File {in_data} not found. Returning.')
-            return ra_samp * np.nan
-        data, hdr = fits.getdata(in_data, header = True)
-
-    else:
-        data = copy.deepcopy(in_data)
-        if not in_hdr is None:
-            hdr = in_hdr
-        else:
-            print(f'{"[ERROR]":<10}', 'Provide a valid header with the data.')
-            return ra_samp*np.nan
-
-    # check dimension of mask
-    dim_data = np.shape(data)
-    if len(dim_data) == 3:
-        is_cube = True
-    else:
-        is_cube = False
-
-    if not is_cube:
-        target_hdr = twod_head(target_hdr)
-
-
-    # Align, if needed
-    if not target_hdr is None:
-        trg_hdr = copy.deepcopy(target_hdr)
-        hdr_out = copy.copy(hdr)
-        wcs_target = WCS(trg_hdr)
-        #need to delete the rest frequency for latest astropy version
-        del hdr_out["RESTF*"]
-        del trg_hdr["RESTF*"]
-        data_out, footprint = reproject_interp((data, hdr_out), trg_hdr, order="nearest-neighbor")
-        data = data_out
-
-    else:
-        print(f'{"[INFO]":<10}', 'No alignment because no target header supplied.')
-
-    #--------------------------------------------------------------
-    #   Sample
-    #--------------------------------------------------------------
-
-    trg_hdr = copy.deepcopy(target_hdr)
-    wcs_target = WCS(trg_hdr)
-    if is_cube:
-        pixel_coords = wcs_target.all_world2pix(np.column_stack((ra_samp, dec_samp, np.zeros(len(dec_samp)))),0)
-    else:
-        pixel_coords = wcs_target.all_world2pix(np.column_stack((ra_samp, dec_samp)),0)
-    samp_x = np.array(np.rint(pixel_coords[:,0]),dtype=int)
-    samp_y = np.array(np.rint(pixel_coords[:,1]),dtype= int)
-
-    n_pts = len(samp_x)
-    dim_data = np.shape(data)
-
-    if is_cube:
-        result = np.zeros((n_pts, dim_data[0]))*np.nan
-    else:
-        result = np.zeros(n_pts)*np.nan
-
-    coverage = np.zeros(n_pts)
-    if is_cube:
-
-        in_map = np.where((samp_x>0)& (samp_x <dim_data[2]) &
-                          (samp_y > 0)& (samp_y < dim_data[1]))[0]
-        in_map_ct = len(in_map)
-    else:
-        in_map = np.where((samp_x>0)& (samp_x <dim_data[1]) &
-                          (samp_y > 0)& (samp_y < dim_data[0]))[0]
-        in_map_ct = len(in_map)
-
-
-    if in_map_ct>0:
-        if is_cube:
-            for kk in range(in_map_ct):
-
-                result[in_map[kk],:] = data[:, samp_y[in_map[kk]],samp_x[in_map[kk]]]
-                coverage[in_map[kk]] =  np.nansum(np.isfinite(data[:, samp_y[in_map[kk]],samp_x[in_map[kk]]]))>=1
-            coverage = np.array(coverage, dtype = int)
-        else:
-            result[in_map] = data[samp_y[in_map],samp_x[in_map]]
-            coverage[in_map] = np.isfinite(data[samp_y[in_map],samp_x[in_map]])
-
-    #--------------------------------------------------------------
-    #   Return
-    #------------------------------------------------------------
-
-    return result, trg_hdr
